Refactor view from `NetworkSelectionView` into `NetworkSelectionRootView`, add `NetworkFilterView`.
