+++ Sources/BraveWallet/Crypto/Stores/NetworkSelectionStore.swift
@@ -7,48 +7,11 @@ import BraveCore
 import BraveShared
 import SwiftUI
 
-struct NetworkPresentation: Equatable, Hashable, Identifiable {
-  enum Network: Equatable, Hashable {
-    case allNetworks
-    case network(BraveWallet.NetworkInfo)
-  }
-  
-  var id: String {
-    switch network {
-    case .allNetworks: return "allNetworks"
-    case let .network(network): return network.id
-    }
-  }
-  let network: Network
-  let subNetworks: [BraveWallet.NetworkInfo]
-  let isPrimaryNetwork: Bool
-  
-  init(
-    network: Network,
-    subNetworks: [BraveWallet.NetworkInfo],
-    isPrimaryNetwork: Bool
-  ) {
-    self.network = network
-    self.subNetworks = subNetworks
-    self.isPrimaryNetwork = isPrimaryNetwork
-  }
-  
-  static let allNetworks: Self = .init(
-    network: .allNetworks,
-    subNetworks: [],
-    isPrimaryNetwork: true
-  )
-}
-
 class NetworkSelectionStore: ObservableObject {
   
   enum Mode: Equatable {
     case select
-    case filter
     case formSelection
-    
-    var isSelectMode: Bool { self == .select }
-    var isFilterMode: Bool { self == .filter }
   }
   
   let mode: Mode
@@ -57,8 +20,6 @@ class NetworkSelectionStore: ObservableObject {
   @Published private(set) var primaryNetworks: [NetworkPresentation] = []
   @Published private(set) var secondaryNetworks: [NetworkPresentation] = []
   
-  /// Network selected to show detail view (test networks)
-  @Published var detailNetwork: NetworkPresentation?
   /// If we are prompting the user to add an account for the `nextNetwork.coin` type
   @Published var isPresentingNextNetworkAlert = false
   /// The network the user wishes to switch to, but does not (yet) have an account for `nextNetwork.coin` type
@@ -77,45 +38,17 @@ class NetworkSelectionStore: ObservableObject {
   }
   
   func update() {
-    let subNetworks: (BraveWallet.NetworkInfo) -> [BraveWallet.NetworkInfo] = { [weak networkStore] network in
-      guard let networkStore = networkStore,
-              WalletConstants.primaryNetworkChainIds.contains(network.chainId),
-              Preferences.Wallet.showTestNetworks.value else {
-        return []
-      }
-      let isPrimaryOrTestnetChainId: (_ chainId: String) -> Bool = { chainId in
-        WalletConstants.primaryNetworkChainIds.contains(chainId)
-        || WalletConstants.supportedTestNetworkChainIds.contains(chainId)
-      }
-      return networkStore.allChains.filter {
-        $0.coin == network.coin
-        && $0.symbol == network.symbol
-        && !networkStore.isCustomChain($0)
-        && isPrimaryOrTestnetChainId($0.chainId)
-      }
-    }
-    
-    var primaryNetworks = networkStore.allChains
-      .filter { WalletConstants.primaryNetworkChainIds.contains($0.chainId) }
+    self.primaryNetworks = networkStore.primaryNetworks
       .map { network in
-        let subNetworks = subNetworks(network)
+        let subNetworks = networkStore.subNetworks(for: network)
         return NetworkPresentation(
           network: .network(network),
           subNetworks: subNetworks.count > 1 ? subNetworks : [],
           isPrimaryNetwork: true
         )
       }
-    if mode == .filter {
-      // users can filter by 'All Networks' but cannot select 'All Networks'
-      primaryNetworks.insert(.allNetworks, at: 0)
-    }
-    self.primaryNetworks = primaryNetworks
 
-    self.secondaryNetworks = networkStore.allChains
-      .filter {
-        !WalletConstants.primaryNetworkChainIds.contains($0.chainId)
-        && !WalletConstants.supportedTestNetworkChainIds.contains($0.chainId)
-      }
+    self.secondaryNetworks = networkStore.secondaryNetworks
       .map { network in
         NetworkPresentation(
           network: .network(network),
@@ -129,7 +62,6 @@ class NetworkSelectionStore: ObservableObject {
     switch mode {
     case .select:
       guard case let .network(network) = network else { return false }
-      detailNetwork = nil
 
       let error = await networkStore.setSelectedChain(network)
       switch error {
@@ -140,14 +72,6 @@ class NetworkSelectionStore: ObservableObject {
       default:
         return true
       }
-    case .filter:
-      switch network {
-      case .allNetworks:
-        networkStore.networkFilter = .allNetworks
-      case let .network(network):
-        networkStore.networkFilter = .network(network)
-      }
-      return true
     case .formSelection:
       switch network {
       case let .network(network):
